Script memebers.php was fixed.
Problem:
Call to a function Sql_Commit_Transaction(). This function must exist or not exist, deppend on
the inclution of adodb.inc library (in the configuration file).
If is not included this library the call fails.

Fix:
Now, was added a control to check if the application are working with AdoDb; only in this case the function is called.
now, if $params isn't an array this parameter is added into an array to avoid the loop failure.